requirements.txt
flask==2.0.1
flask-oauthlib==0.9.6
flask-session==0.4.0
xero-python==1.19.0
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
pandas==1.5.3
python-dateutil==2.8.2